Latest core changes from Main

+++ Source/Core/Core/Src/HW/WiimoteReal/IONix.cpp
@@ -71,7 +71,7 @@ void WiimoteScanner::FindWiimotes(std::vector<Wiimote*> & found_wiimotes, Wiimot
 	inquiry_info scan_infos[max_infos] = {};
 	auto* scan_infos_ptr = scan_infos;
 	found_board = NULL;
-	
+
 	// Scan for bluetooth devices
 	int const found_devices = hci_inquiry(device_id, wait_len, max_infos, NULL, &scan_infos_ptr, IREQ_CACHE_FLUSH);
 	if (found_devices < 0)
@@ -86,7 +86,7 @@ void WiimoteScanner::FindWiimotes(std::vector<Wiimote*> & found_wiimotes, Wiimot
 	for (int i = 0; i < found_devices; ++i)
 	{
 		ERROR_LOG(WIIMOTE, "found a device...");
-		
+
 		// BT names are a maximum of 248 bytes apparently
 		char name[255] = {};
 		if (hci_read_remote_name(device_sock, &scan_infos[i].bdaddr, sizeof(name), name, 1000) < 0)
@@ -263,9 +263,9 @@ int Wiimote::IORead(u8* buf)
 	return r;
 }
 
-int Wiimote::IOWrite(u8 const* buf, int len)
+int Wiimote::IOWrite(u8 const* buf, size_t len)
 {
-	return write(int_sock, buf, len);
+	return write(int_sock, buf, (int)len);
 }
 
 }; // WiimoteReal
